Add logJSFunctionExecution in CE
