+++ app/client/src/ce/sagas/JSFunctionExecutionSaga.ts
@@ -1,0 +1,3 @@
+export function* logJSFunctionExecution(data: any) {
+  return data;
+}
